{"schema":"decodex/commit/1","summary":"Move overlay redraw runtime","authority":"manual"}
